Fixed pushBlock function

+++ Zumo32U4/robot.hpp
@@ -207,7 +207,7 @@ public:
    * 
    * @param[in] delayMillis The amount of time for the movement in milliseconds
    */
-  void timeMove(unsigned int time);
+  void timeMove(uint16_t time);
 
   /**
    * Pushes the block forwards by a given amount of cells
@@ -222,7 +222,7 @@ private:
   uint16_t _centeringDelay;           // The amount of milliseconds to wait after finding an intersection. This is needed to center the robot on the cross
   uint8_t _pathSeekCompensation;      // The initial number of degrees to rotate when the robot is searching the lost black line. See fixPath function
   uint8_t _speedCompensation;         // The speed increase used to make the robot slightly rotate when it arrives at an intersection but it is not parallel to it
-  uint8_t _blockCenteringDelay;       // The amount of milliseconds do wait after finding an intersection. This is needed to center the block on the cross after pushing it
+  uint16_t _blockCenteringDelay;       // The amount of milliseconds do wait after finding an intersection. This is needed to center the block on the cross after pushing it
   matrix<cell_content>* _grid;        // A pointer to the grid of cells. This is used to avoid the need to pass it as a parameter to most movement functions
   enum object_movement _orientation;  // The direction that the robot is facing
